StatsController for time-to-data-arrival and other things (#246)
Initial metric is time-to-data-arrival on the catalog page. For now, the controller uses `window.location.hash` when it's instantiated to determine whether or not the statistic should apply. This is important because there is no navigation+reload when moving between front page, catalog page, and metrics page. The metric is only reported when the user first loads the page looking at a catalog (and non-empty catalog data eventually arrive).