Allow deploying without symlinking node_modules/foam2
The exact version of foam2 deployed is saved in package-lock.json and
not commiting any change results in "Your working tree is not clean",
so since we don't need to change the version of foam2 used without
also updating the confluence repo, deploy can be simplified.

This is the only (apparent) remaining reason that node_modules/foam2
needs to be symlinked to a full foam2 checkout.