Upgrade to Babel 7
https://babeljs.io/docs/en/v7-migration

There's no need to run babel at all when running tests or with a local
dev server, so drop `ES6_LOADER_OPTIONS_DEV` entirely.

The use of `comments: false` in the babel options makes no difference
at all to the size of the bundles, presumably because comments are
stripped again by the minimizer, so that option is not preserved.