Don't use babel when running tests using karma (#398)
Currently instrumentation is running after babel, which results in
incorrect and nonsensical coverage information. The node vs. web
coverage here is an example:
https://codecov.io/gh/GoogleChromeLabs/confluence/src/f6802df5a6be89042a1db1e7b82f0a0a90906caf/lib/dao/indexed_dao.es6.js

This could be fixed by changing the order:
https://github.com/GoogleChromeLabs/confluence/pull/404

However, it's not clear if that's correct:
https://github.com/webpack-contrib/istanbul-instrumenter-loader/issues/86

There's no need to run Babel to test stable versions of Chrome or Firefox,
so just skip babel-loader.